Struct fluvio_experimental::Partition

pub struct Partition { /* fields omitted */ }

A Partition is a unit of streaming that enables messages to be sent in parallel.

You can think of a partition as a fraction of a Topic. Every message that is produced to a topic will be assigned to a partition, which is responsible for streaming and storing that message. The thing that makes partitions special is that separate partitions may be processed by entirely different machines. This ultimately allows you to increase the potential throughput of your topic.

+-------+-----------+
|       | Partition |
| Topic |-----------+
|       | Partition |
+-------+-----------+

Having more than one partition for a topic changes the ordering semantics of your topic. All of the messages belonging to a single partition will arrive and be stored in order. However, there is no ordering guarantee for messages that belong to different partitions.

Example

Suppose you have a service for scanning admissions tickets at sporting events. Your system has a barcode scanner that reads a number from each ticket and produces them to a topic called admissions. Lets say this topic has two partitions, and that you send all of the even ticket numbers to one partition and all of the odd ticket numbers to the other partition.

+--------------------+----------------------------+
|                    | Partition 0 (even tickets) |
| Topic (admissions) |----------------------------+
|                    | Partition 1 (odd tickets)  |
+--------------------+----------------------------+

The code to produce ticket numbers from your barcode scanner to the admissions topic might look something like this:

let mut even_partition = admissions_topic.partition(0)?;
let mut odd_partition = admissions_topic.partition(1)?;

while let Some(ticket_number) = barcode_scanner.next().await {
    if ticket_number % 2 == 0 {
        even_partition.produce_string_message(format!("{}", ticket_number)).await?;
    } else {
        odd_partition.produce_string_message(format!("{}", ticket_number)).await?;
    }
}

Game night arrives, and your first few fans in line have the following ticket numbers

12, 13, 55, 89, 90, 44, 23

Those ticket numbers would be sent to the admissions topic like this

+--------------------+----------------------------+
|                    | Partition 0 (even tickets) |
|                    | [ 12, 90, 44 ]             |
| Topic (admissions) |----------------------------+
|                    | Partition 1 (odd tickets)  |
|                    | [ 13, 55, 89, 23 ]         |
+--------------------+----------------------------+

Notice that all of the even-numbered tickets are stored in the order that they arrived in, and that the same is true for the odd-numbered tickets. However, it is important to realize that there is no longer any ordering guarantee between even and odd tickets. It would be impossible to tell whether ticket 12 arrived before or after ticket 89.

Implementations

impl Partition

pub async fn fetch_string_messages<'_, R: RangeBounds<i64>>(
&'_ self,
range: R
) -> Result<Vec<String>, FluvioError>

pub async fn produce_string_message<'_, S: Into<String>>(
&'_ mut self,
message: S
) -> Result<(), FluvioError>

Produces a string message to this partition